Replace timestamp by bevy time and few refactors
