+++ crates/bevy_asset_preview/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2024"
 
 [dependencies]
 bevy = { workspace = true, features = ["webp"] }
-image = { version = "0.25", features = ["webp"] }
+image = "0.25"
 
 [dev-dependencies]
 tempfile = "3"
